#1626: remove no longer needed code handling local identifiers
This is a partial revert of 03ce64e15195cd03. That functionality
is no longer needed in SDK as it was moved to tiger in #906.

On top of the revert there are changes in ObjRefConverter,
namely the code handling identifiers in measure and ranking
filters is now merged as it is just the same.